Introducing: The Visual Aid®
Say goodbye to our old scripts, and say hello to the new and improved, it’s
bigger, it’s badder, ladies and gentlemen, it’s too much for
Mr. Incredible!

The one script to rule them all. Basically it will listen to all the
things, check it against some regexes and look up a random url from a
json file.

Why? I think this would make it easier to contribute to the file, no
need to create a whole script, just prop in a regex and a file and
you're good to go.

Different opinions are available, so up for debate whether this should
go in or not.

Thoughts?